# Conflicts:
#	common/src/main/java/org/geogebra/common/GeoGebraConstants.java

+++ common/src/main/java/org/apache/commons/math3/stat/inference/MannWhitneyUTest.java
@@ -141,7 +141,7 @@ public class MannWhitneyUTest {
          * U1 = R1 - (n1 * (n1 + 1)) / 2 where R1 is sum of ranks for sample 1,
          * e.g. x, n1 is the number of observations in sample 1.
          */
-        final double U1 = sumRankX - ((long) x.length * (x.length + 1)) / 2;
+        final double U1 = sumRankX - ((double) x.length * (x.length + 1)) / 2;
 
         /*
          * It can be shown that U1 + U2 = n1 * n2
